# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Admin Platform API
  version: "1.0.0"
  description: |
    Backend API for the Admin Platform.
    All module endpoints require **Bearer JWT** authentication and pass through a
    four-layer guard chain: `authenticateJwt -> requireRole -> requireFeatureEnabled -> requirePermission`.

    **Error envelope** (all non-2xx responses):
    ```json
    { "code": "<ErrorCode>", "message": "<human string>" }
    ```

    **Pagination envelope** (list endpoints):
    ```json
    { "items": [...], "page": 1, "limit": 50, "total": 123 }
    ```
  contact:
    name: Admin Platform
servers:
  - url: http://localhost:7002
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Two-tier JWT.  Tokens are signed with role-specific secrets
        (`JWT_SECRET_SUPER_ADMIN` / `JWT_SECRET_ADMIN`).
        Payload: `{ sub, email, role }`.

  # ─── Reusable schemas ───────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }

    # ── Auth ──
...
    MailboxMessage:
      type: object
      properties:
        _id: { type: string }
        subject: { type: string }
        body: { type: string }
        fromAddress: { type: string }
        fromName: { type: string }
        toAddresses: { type: array, items: { type: string } }
        ccAddresses: { type: array, items: { type: string } }
        bccAddresses: { type: array, items: { type: string } }
        folder: { type: string, enum: [inbox, sent, drafts, trash, archive] }
        status: { type: string, enum: [unread, read, flagged] }
        isStarred: { type: boolean }
        sentAt: { type: string, format: date-time }
        ownerUserId: { type: string }
        inReplyToId: { type: string }
        createdAt: { type: string, format: date-time }

    # ── Projects ──
...
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: MongoDB ObjectId

  # ─── Reusable responses ───────────────────────────────────────────
...
paths:
...
  /api/v1/mailbox/messages:
    get:
      tags: [Mailbox]
      summary: List messages (owner-scoped)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - { name: folder, in: query, schema: { type: string, enum: [inbox, sent, drafts, trash, archive], default: inbox } }
        - { name: allUsers, in: query, schema: { type: boolean }, description: super_admin only }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationMeta"
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/MailboxMessage" } }
    post:
      tags: [Mailbox]
      summary: Compose message (rate-limited 60/min)
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailboxMessage" }
  /api/v1/mailbox/messages/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [Mailbox]
      summary: Get message (auto-marks inbox as read)
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailboxMessage" }
    patch:
      tags: [Mailbox]
      summary: Update draft message
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailboxMessage" }
    delete:
      tags: [Mailbox]
      summary: Delete (trash first, then permanent)
      responses:
        "204": { description: Deleted }
  /api/v1/mailbox/messages/{id}/send:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    post:
      tags: [Mailbox]
      summary: Send draft
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailboxMessage" }
  /api/v1/mailbox/messages/{id}/move:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    post:
      tags: [Mailbox]
      summary: Move to folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [folder]
              properties:
                folder: { type: string, enum: [inbox, sent, drafts, trash, archive] }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailboxMessage" }
  /api/v1/mailbox/messages/{id}/star:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    post:
      tags: [Mailbox]
      summary: Toggle star
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailboxMessage" }
  /api/v1/mailbox/messages/{id}/read:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    post:
      tags: [Mailbox]
      summary: Mark as read
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MailboxMessage" }
...